requirements.txt
# ============================================
# Solubility Oracle - Python Dependencies
# ============================================
# Install: pip install -r requirements.txt
# ============================================

# ----- Core Deep Learning -----
torch>=2.0.0
torch-geometric>=2.3.0
# intel-extension-for-pytorch>=2.0.0  # Optional: Intel GPU acceleration

# ----- Chemistry / Molecular Processing -----
rdkit>=2023.03.1
mordred>=1.2.0

# ----- Data & ML -----
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# ----- Visualization -----
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# ----- Web Framework -----
streamlit>=1.28.0
fastapi>=0.103.0
uvicorn>=0.23.0

# ----- Experiment Tracking -----
mlflow>=2.7.0
optuna>=3.3.0

# ----- Utilities -----
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.66.0
requests>=2.31.0

# ----- Development & Testing -----
pytest>=7.4.0
# black>=23.9.0
# isort>=5.12.0
# mypy>=1.5.0

